Coreboot Support
Cabability added to the STM to allow it to boot 32 bit SMI Handlers (along with 64-bit SMI handers) for Coreboot.
Other additions:
Corrected problems induced by not flushing the caches when exiting SMM.  In certain cases parts of data structures would not be stored.
Enhanced some debug statemente for clarity