Coreboot support and bug fixes
(1) Added coreboot support.  Mainly, allowed the SMI Handler to be started as a
    32-bit virtual machine.

(2) Fixed a problem during STM initialization where the relocation tables
    could not be located because of differences in compiler behaviors
    in handling the "&"

(3) Renamed Readme.Newbuild to Readme>STMPE and updated it with additional
    CMAKE build options to provide building for coreboot, alternative
    UART ports, and resizing of STM Heap memory.

(4) Fixed a bug in AsmVmLaunch and AsmVmResume where the rsi and rdi registers
    were swapped when restoring the system state

(5) Some reformatting of comments and code.

(6) Added additional debug support functions to dump VM and STM memory locations.

(7) Added the VM number when VMCS registers are dumped.